remove game dependency from hud

+++ src/ui/hud/textfieldhudgetvoxels.cpp
@@ -68,7 +68,7 @@ float TextFieldHudgetVoxels::scale() {
 }
 
 const glm::vec3 TextFieldHudgetVoxels::offsetToCenter(bool upper, bool left) const {
-    float horizontalOffset = left ? m_text.length() : 0;
+    float horizontalOffset = left ? static_cast<float>(m_text.length()) : 0;
     float verticalOffset = upper ? m_height : -m_height;
 
     return glm::vec3(m_offset + m_width * horizontalOffset - m_width / 2, verticalOffset, 0);
